One more QList-based warning bites the dust
I think this is the last of those (haven't looked in the plugins)

+++ mythtv/programs/mythbackend/mainserver.cpp
@@ -2806,7 +2806,7 @@ void MainServer::HandleQueryFreeSpaceSummary(PlaybackSock *pbs)
     }
     else
     {
-        unsigned int index = strlist.size() - 2;
+        unsigned int index = (uint)(strlist.size()) - 2;
         shortlist << strlist[index++];
         shortlist << strlist[index++];
     }
